relocate preloaded client and rename crate from hokulea-witgen-client to hokulea-witgen
