+++ Cargo.toml
@@ -36,7 +36,7 @@ hokulea-client = { path = "crates/client", version = "0.1.0", default-features =
 hokulea-eigenda = { path = "crates/eigenda", version = "0.1.0", default-features = false }
 hokulea-proof = { path = "crates/proof", version = "0.1.0", default-features = false }
 hokulea-compute-proof = { path = "crates/compute-proof", version = "0.1.0", default-features = false }
-hokulea-witgen-client = { path = "crates/witgen-client", version = "0.1.0", default-features = false }
+hokulea-witgen = { path = "crates/witgen", version = "0.1.0", default-features = false }
 
 # EigenDA v2 struct
 # TODO: this should be moved elsewhere, probably to https://github.com/Layr-Labs/eigenda-rs
