+++ bin/host/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2021"
 hokulea-proof.workspace = true
 hokulea-client-bin.workspace = true
 hokulea-eigenda.workspace = true
-hokulea-witgen-client.workspace = true
 
 # Kona
 kona-preimage = { workspace = true, features = ["std"] }
