--- crates/witgen-client/src/lib.rs
@@ -1,2 +1,0 @@
-pub mod witgen_client;
-pub mod witness;
