+++ crates/witgen/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "hokulea-witgen-client"
+name = "hokulea-witgen"
 version = "0.1.0"
 edition = "2021"
