+++ crates/witgen/src/witness.rs
